rdar://16833552 (Add a module map file for PrivateFrameworks/WebKitLegacy)
Define modules for the Cocoa Touch environment. macOS can't define modules in WebKitLegacy because it's a nested framework on that platform. The public module is empty because there are no public headers. The private module map sets PrivateHeaders as the umbrella directory because there isn't an umbrella header already. Then it makes an explicit submodule for each header so as to mimic the non-modular environment.

Enable the modules verifier tool for Apple Internal (it's not supported in the public Xcode).

Add the Apple Internal guards for NSURLDownload from WebDownload.h to NSURLDownloadSPI.h so it can be included in the module.

WebCreateFragmentInternal.h and WebGeolocationCoreLocationProvider.h are Objective-C++ headers, but the WebKitLegacy module should be usable by plain Objective-C clients. Add C++ guards so that it can be included in the module.

Switch a few quoted includes to framework style angle includes, which are required for modules.

Canonical link: https://commits.webkit.org/242465.17@safari-613.1.4-branch
git-svn-id: https://svn.webkit.org/repository/webkit/branches/safari-613.1.4-branch@284023 268f45cc-cd09-0410-ab3c-d52691b4dbfc